# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Xoycoin Exchange",
    "description": "Métodos de la API de Xoycoin Exchange",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-xoycoin.xellcoders.com",
      "description": "Xoycoin Exchange API"
    }
  ],
...
  "paths": {
...
    "/webhooks/": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Create webhook",
        "description": "Crea un webhook.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/webhooks_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Webhook object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Webhook"
                },
                "example": {
                  "id": "dd157249-9729-4ce7-a05e-31c765fe323b",
                  "url": "http://www.domain.com/webhook",
                  "contentType": "application/json",
                  "status": "ACTIVE"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/webhooks/event/": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Assign webhook event",
        "description": "Asigna un evento a un webhook.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/webhooks_event_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/webhooks/{webhookId}": {
      "get": {
        "tags": [
          "webhooks"
        ],
        "summary": "Get webhook by id",
        "description": "Obtiene un webhook mediante el id proporcionado.",
        "parameters": [
          {
            "name": "webhookId",
            "in": "path",
            "description": "Id of the webhook to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Webhook"
                },
                "example": {
                  "id": "dd157249-9729-4ce7-a05e-31c765fe323b",
                  "url": "http://www.domain.com/webhook",
                  "contentType": "application/json",
                  "status": "ACTIVE"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Webhook not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/webhooks/company/{companyId}": {
      "get": {
        "tags": [
          "webhooks"
        ],
        "summary": "Get webhook by company id",
        "description": "Obtiene un webhook mediante el id de una compañia.",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "description": "Id of the company to return webhooks",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhooks array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Webhook"
                },
                "example": [
                  {
                    "id": "dd157249-9729-4ce7-a05e-31c765fe323b",
                    "url": "http://www.domain.com/webhook",
                    "contentType": "application/json",
                    "status": "ACTIVE"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Cannot get webhooks"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/webhooks/activate": {
      "put": {
        "tags": [
          "webhooks"
        ],
        "summary": "Activate webhook",
        "description": "Activa un webhook.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/webhooks_activate_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/webhooks/inactivate": {
      "put": {
        "tags": [
          "webhooks"
        ],
        "summary": "Inactivate webhook",
        "description": "Inactiva un webhook.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/webhooks_inactivate_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      },
...
      "webhooks_body": {
        "required": [
          "contentType",
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      },
      "webhooks_event_body": {
        "required": [
          "eventId",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "eventId": {
            "type": "integer"
          }
        }
      },
      "webhooks_activate_body": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "webhooks_inactivate_body": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}